requirements.txt
pymodm==0.4.1
djangorestframework==3.8.2
django==2.1
markdown==2.6.11
django-extensions==2.1.0
git+git://github.com/lokoArt/pymodm_rest.git
mysqlclient==1.3.13
protobuf==3.6.1
coreapi==2.3.3
django-storages==1.6.6
zappa==0.46.2
